This commit deals with two considerations when storing our business logic.
It is possible to do regular snapshots of mobx-keystone and store those in the storage but this system
would result in storing very big objects very often, but alternatively, mobx-keystone allows to keep
track of any changes using `onPatches`.

Each patch is a small json object that can be stored and later combined into a bigger object.

The storage in this commit makes it also possible to define "indices" on the storage that give access to combined versions which
can be stored as snapshots at any given time.

The implementation further allows any data to be stored in a regular filesystem. This is important
because the SecretStore of android (and expo) is limited to a 2MB. This means entries can be stored
in encrypted fashion on the hard disk, encrypted without running into limitations.