Fixing persistence of key on device.
A missing update and inverse boolean condition flipped the persistence of a vault key on the device. Before this commit a key on the device would have been persisted if a lockee is added but never removed - triggering the whole system ineffective. Now the key is persisted on the device as long as no lockee is added but the persistence is lifted once a lockee is added.